fix: disable auto-collapse (#219)
Docusaurus isn't build to handle the same pages being added to the sidebar, so when auto-collapse is on, no matter the page you are clicking, it select the last one, which is an unwanted behaviour. When it's off, it select the page everywhere, which is also unwanted, but less a problem as it stick to the page you choose at least.