ci: wait for previous release action before continuing
When a PR is merged to main, while another one was merged and not deployed on production, a Helm chart issue may arise. Unfortunately, the concurrency option from GitHub Actions cancel previous pending actions, which is not the intented result we want (there is no way yet to change this behaviour https://github.com/orgs/community/discussions/41518). It's usually not a problem as new PR needs to be updated with main content, but it may prevent a deployment is the new PR isn't on the docs: the new release action wouldn't run, and the old one be cancelled, you the previous PR changes to the docs wouldn't be published until the next PR, or manual run of the action. Trying this waiting action as if it's working well, it could solve the issue.