chore: add .env for Typesense manual indexing
Since I have to manually index the content for the TypeSence search for now, adding back this .env variable which is needed. The only secret is not added here. Just making my life easier as I always delete it when resetting the git repo, and forget the values I need.