docs: first draft of K3s docs
The diagrams are the ones from Civo: will update in another PR as I have to do some SVG changes first.. I also removed GitHub where it was possible to comment the option (not easily on common pages). Laslty, it's missing the deprovision information as I need to be able to test K3s first to document properly the process.