feat: remove Docusaurus trailing slash
It confirmed my hypothesis on the crawler issue on our docs. To fix it, I need to update how ngnix serve links, as using this Docusaurus option cause a lot of issues with relative path, so it's not the way to go. There is always an issue with the crawler not scraping redirections it detected, so I'll create an issue and check with TypeSense, but in the meantime, it's easy to fix it on our side with our deployment since we do not need trailing slashes on URLs.